Extracted the get number lenght in new module.
