+++ src/long-multiplication-command-line/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod breakdown;
 pub mod generate;
+pub mod length;
 pub mod multiplication;
